Refactor - 파일 통합 & View 수정
- Extension 파일 통합
- View padding 값 수정
