--- Baggle/Baggle/DesignSystem/Extension/Color+Temp.swift
@@ -1,10 +1,0 @@
-//
-//  Color+Temp.swift
-//  promiseDemo
-//
-//  Created by youtak on 2023/07/17.
-//
-
-// 빈 파일입니다.
-
-// Extension을 사용해 Color와 font, 이미지를 관리합니다.
